Switch to creating the  user interface with Qt-Creator
Switched to creating the user interface with Qt-Creator. In the process
I will rewrite parts of the code as I redo the overall application
design. The current state of the code is in the middle of the redesgin.

Programm runs when build with qmake. CMakeLists.txt is not yet updated
so cmake will not build the project.