Progress with the graph painting
Programm now paints a graph based on the values it gets from
smbtamonitor-gen via the socket. The graph has the known issues already
present in the old versions of smbtamonitor-qt, but we
do now have a realtime graph that reflects the traffic caused by
tortureserv. Also there are still some other bugs with that, but those will be
dealt with in the next days.