Graph now moves from left to right
Graph now moves from the left to the right side, as requestet. When the
right side is hit the newest value will stay on the right side, the
oldest values will disappear on the left side.

The graph is currently running on speed for the ease of testing.
Interpolation will be implemented next.