Reduced code redundancy and improved rescaling
Moved most values that are used to compute the graph to class-wide
variables. This makes changing easier, especially for runtime
adjustments. The number of stored values and the numer of displayed
values are now handled differently, which allow to implement history
functionality in the future.

Improved the rescaling method. If a graph maximum value now moves out of
the displayed range, the graph will be rescaled according to the
remaining highest value, including axes etc...